Upgrade libs

+++ CarHomeWidget/src/main/java/com/anod/car/home/prefs/colorpicker/AlphaPatternDrawable.java
@@ -72,8 +72,8 @@ public class AlphaPatternDrawable extends Drawable {
         int height = bounds.height();
         int width = bounds.width();
 
-        numRectanglesHorizontal = (int) Math.ceil((width / mRectangleSize));
-        numRectanglesVertical = (int) Math.ceil(height / mRectangleSize);
+        numRectanglesHorizontal = (int) Math.ceil((width / (float)mRectangleSize));
+        numRectanglesVertical = (int) Math.ceil(height / (float)mRectangleSize);
 
         generatePatternBitmap();
